requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
dataclasses-json==0.6.1
Django==4.2.5
django-storages==1.14
djangorestframework==3.14.0
dnspython==2.4.2
frozenlist==1.4.0
greenlet==2.0.2
idna==3.4
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.302
langsmith==0.0.40
loguru==0.7.2
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
nltk==3.8.1
numexpr==2.8.7
numpy==1.26.0
openai==0.28.1
packaging==23.1
pinecone-client==2.2.4
psycopg2==2.9.7
pydantic==2.4.1
pydantic_core==2.10.1
PyPDF2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
scipy==1.11.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.21
sqlparse==0.4.4
tenacity==8.2.3
tiktoken==0.5.1
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
urllib3==2.0.5
whitenoise==6.5.0
yarl==1.9.2
